requirements.txt
asn1crypto==0.23.0
bcrypt==3.1.4
blinker==1.4
certifi==2017.7.27.1
cffi==1.11.2
chardet==3.0.4
click==6.7
crcmod==1.7
cryptography==2.3.1
Faker==4.1.1
Flask==1.0.2
Flask-Cors==3.0.8
Flask-Docs==0.1.3
Flask-Login==0.4.0
Flask-Mail==0.9.1
Flask-OAuthlib==0.9.4
Flask-SQLAlchemy
Flask-Script==2.0.6
Flask-Session==0.3.1
Flask-WTF==0.14.2
geoip2==2.6.0
idna==2.6
ipaddress==1.0.18
itsdangerous==0.24
Jinja2==2.10
lxml==4.1.1
Markdown==2.6.11
MarkupSafe==1.0
maxminddb==1.3.0
oauthlib==2.0.6
olefile==0.44
paramiko==2.4.2
Pillow==4.3.0
psutil==5.4.0
pyasn1==0.3.7
pycparser==2.18
pycrypto==2.6.1
pycryptodome==3.6.6
PyJWT==1.5.3
PyNaCl==1.2.0
pytz==2017.3
PyYAML==3.12
redis==2.10.6
regex==2017.9.23
requests==2.20.1
requests-oauthlib==0.8.0
six==1.11.0
tld==0.7.9
urllib3==1.24
Werkzeug==0.14.1
WTForms==2.1
Flask-Migrate
PyMySQL
shortuuid
flask_paginate
